requirements.txt
# AI Movie Shorts Pipeline - Dependencies

# Gemini API (new SDK — replaces deprecated google-generativeai)
google-genai>=1.0.0

# Kokoro TTS (free local voice synthesis)
kokoro>=0.9.4
soundfile>=0.12.1

# Playwright (browser automation for CinemagicHD web app)
playwright>=1.44.0

# YouTube upload
google-api-python-client>=2.120.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# Subtitle fetching
requests>=2.31.0

# Video/audio processing helpers
moviepy>=1.0.3
pydub>=0.25.1

# Utilities
python-dotenv>=1.0.0
rich>=13.7.0
schedule>=1.2.1
